Switches LineSurface to BigDecimal (#18)
This change swaps out Double with BigDecimal in the Point and Slope variables. Interestingly,
this made the Floor adjuster even MORE picky, because now we have more decimal information
but with the same problem of not being able to represent a point 100%. When comparing the
Mobile's Position with the Line's equation, there was only a one billionth difference, but
that still returned false.

The solution is to round these numbers to ~5 decimal places before comparing. This seems
to work well enough, but I'm still afraid that there's a loss of data with each iteration.
Now I understand how I could sometimes fall out of the map in Quake.

+++ app/game/mobile/MobileModule.scala
@@ -12,8 +12,8 @@ import game.util.math.LineModule
 trait MobileModule extends EventModule with LineModule {
   this: RoomModule ⇒
 
-  case class Movement( val x: Double, val y: Double )
-  case class Position( val x: Double, val y: Double ) extends PointLike {
+  case class Movement( val x: BigDecimal, val y: BigDecimal )
+  case class Position( val x: BigDecimal, val y: BigDecimal ) extends PointLike {
     lazy val head = Point( x, y + 2 )
     lazy val feet = Point( x, y - 2 )
     lazy val right = Point( x + 1, y )
